+++ client/src/pages/HomePage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Dimensions, Text, View } from 'react-native';
 import ActionButton from '../components/ActionButton';
-import AddExpense from '../modals/AddExpense';
+import ExpenseForm from '../modals/ExpenseForm';
 
 const HomePage = () => {
   return (
@@ -11,7 +11,7 @@ const HomePage = () => {
         <Text>This is the home page</Text>
       </View>
       <ActionButton height={Dimensions.get('window').height}>
-        <AddExpense />
+        <ExpenseForm />
       </ActionButton>
     </>
   );
